requirements.txt
argcomplete==1.9.3
asgiref==3.2.7
asn1crypto==1.3.0
base58==0.2.4
#bleach==5.0.1
bleach==5.0.1
bootstrap-admin==0.4.0
boto==2.49.0
boto3==1.24.89
botocore==1.27.89
cachetools==3.1.0
certifi==2020.4.5.1
cffi==1.15.1
cfn-flip==1.1.0.post1
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
#cryptography==2.6.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.5
django-ckeditor==5.6.1
django-cors-headers==2.4.1
django-debug-toolbar==1.11
django-extensions==2.1.6
django-filter==2.2.0
django-import-export==2.0.2
django-js-asset==1.2.1
django-mathfilters==0.4.0
django-private-storage==2.2
django-request==1.5.5
django-rest-swagger==2.2.0
django-storages==1.6.6
django-webpack-loader==0.6.0
djangorestframework==3.9.1
docutils==0.14
#drf-api-logger == 1.1.11
durationpy==0.5
et-xmlfile==1.0.1
facebook-sdk==3.1.0
future==0.16.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==20.0.4
hjson==3.0.1
httplib2==0.12.1
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
jmespath==0.9.3
jxmlease==1.0.3
kappa==0.6.0
lambda-packages==0.20.0
#lxml==4.3.3
lxml==4.9.1
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==2.0.3
newrelic==7.2.4.171
#numpy==1.16.3
numpy==1.20.3
oauth2client==4.1.3
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.3
#pandas==1.5.0
pandas==1.4.4
Pillow==8.2.0
placebo==0.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.2
python-docx==0.8.10
python-slugify==1.2.4
pytz==2022.4
PyYAML==6.0
#PyYAML==3.13
razorpay==1.2.0
requests== 2.28.1
rsa==4.0
s3transfer==0.6.0
#setuptools==63.3.0
setuptools==63.3.0
simplejson==3.17.0
six==1.12.0
sqlparse==0.2.4
tablib==1.1.0
toml==0.10.0
tqdm==4.19.1
#troposphere==2.4.1
troposphere==4.1.0
#"arn":"aws:lambda:ap-south-1:721772295061:function:erdr-development"
#"lambda_description": "aws:states:opt-out",
#"Resource" : "arn:aws:lambda:ap-south-1:721772295061:function:erdr-development"
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.26.12
#webencodings==0.5.1
webencodings==0.5.1
Werkzeug==0.14.1
#wheel==0.37.1
wheel==0.37.1
wsgi-request-logger==0.4.6
xlrd==1.2.0
xlwt==1.3.0
zappa==0.55.0
azure-storage-blob==12.16.0
azure-common==1.1.28
azure-core==1.26.4
azure-nspkg==3.0.2
azure-storage==0.20.3